Sandboxes lib file, adjust imports, sync messaging
